+++ packages/plugin-gallery/web/src/consts.ts
@@ -1,0 +1,5 @@
+export const galleryModals = {
+  settings: 'gallerySettings',
+};
+
+export const fileInputAccept = '.jpg,.png,.gif,.jpeg,.jpe,.jfif,.bmp,.heic,.heif,.tfif,.tif,.webp';