+++ packages/toolbars-ui/web/src/buttons/SettingsButton/SettingsButton.tsx
@@ -4,6 +4,7 @@ import { ToggleButton } from '../../components';
 import { RicosContext, ModalContext } from 'ricos-context';
 import { SettingsIcon } from '../../icons';
 import type { IToolbarItem } from 'ricos-types';
+import { UploadServiceContext } from 'wix-rich-content-common';
 
 interface Props {
   toolbarItem: IToolbarItem;
@@ -12,14 +13,14 @@ interface Props {
 const SettingsButton: FC<Props> = ({ toolbarItem }) => {
   const modalService = useContext(ModalContext) || {};
   const { isMobile, t } = useContext(RicosContext) || {};
-
+  const { uploadService, updateService } = useContext(UploadServiceContext);
   const onClick = toolbarItem.commands.click;
   const node = toolbarItem.attributes.selectedNode;
 
   return (
     <ToggleButton
       Icon={SettingsIcon}
-      onClick={() => onClick({ modalService, isMobile, node })}
+      onClick={() => onClick({ modalService, isMobile, node, uploadService, updateService })}
       dataHook="baseToolbarButton_settings"
       tooltip={t('SettingsButton_Tooltip')}
     />